+++ controlgallery/AtomUIGallery/ShowCases/DataDisplay/TreeView/ViewModels/TreeViewViewModel.cs
@@ -96,6 +96,20 @@ public class TreeViewViewModel : ReactiveObject, IRoutableViewModel
         set => this.RaiseAndSetIfChanged(ref _customizeCollapseExpandTreeDefaultExpandedPaths, value);
     }
 
+    private IList<TreeNodePath>? _semanticPartTreeDefaultSelectedPaths;
+    public IList<TreeNodePath>? SemanticPartTreeDefaultSelectedPaths
+    {
+        get => _semanticPartTreeDefaultSelectedPaths;
+        set => this.RaiseAndSetIfChanged(ref _semanticPartTreeDefaultSelectedPaths, value);
+    }
+
+    private IList<TreeNodePath>? _semanticPartPreviewDefaultSelectedPaths;
+    public IList<TreeNodePath>? SemanticPartPreviewDefaultSelectedPaths
+    {
+        get => _semanticPartPreviewDefaultSelectedPaths;
+        set => this.RaiseAndSetIfChanged(ref _semanticPartPreviewDefaultSelectedPaths, value);
+    }
+
     private IList<ITreeItemNode>? _asyncLoadTreeNodes;
     public IList<ITreeItemNode>? AsyncLoadTreeNodes
     {